pyproject.toml
[project]
name = "django-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Angelo Tobias",email = "tobias.danielangelo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "djangorestframework-camel-case (>=1.4.2,<2.0.0)",
    "django-rest-knox (>=5.0.2,<6.0.0)",
    "inflect (>=7.5.0,<8.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "lzstring (>=1.0.4,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
